all issue fixed
