+++ src/components/ComplaintCard/ComplaintCard.tsx
@@ -1,7 +1,6 @@
 import type React from "react"
 import "./ComplaintCard.css"
 import { Complaint } from "../../data/types"
-// import { formatRelativeTime } from "../../data/utils"
 
 
 interface ComplaintCardProps {
@@ -11,8 +10,6 @@ interface ComplaintCardProps {
 
 const ComplaintCard: React.FC<ComplaintCardProps> = ({ complaint, animationDelay = 0 }) => {
   const { Title, Body } = complaint
-  // const { Title, Body, CreatedAt } = complaint
-  // const relativeTime = formatRelativeTime(CreatedAt)
   const hasTitle = Boolean(Title.trim())
   const hasBody = Boolean(Body.trim())
 
@@ -29,7 +26,6 @@ const ComplaintCard: React.FC<ComplaintCardProps> = ({ complaint, animationDelay
         ) : (
           <h3 className="complaint-title missing">No Title</h3>
         )}
-        {/* <span className="complaint-time">{relativeTime}</span> */}
       </div>
 
       <div className="complaint-card-body">
