+++ src/components/ComplaintForm/ComplaintForm.css
@@ -4,6 +4,7 @@
   padding: var(--spacing-xl);
   transition: transform var(--transition-normal),
     box-shadow var(--transition-normal);
+  width: 100%;
 }
 
 .complaint-form-container h2 {
@@ -58,11 +59,7 @@
 }
 
 .submit-button {
-  background: linear-gradient(
-    135deg,
-    var(--primary-color),
-    var(--secondary-color)
-  );
+  background: linear-gradient(90deg, var(--accent-color), var(--primary-color));
   color: white;
   border: none;
   border-radius: var(--border-radius-md);
@@ -91,7 +88,7 @@
     rgba(255, 255, 255, 0.2),
     transparent
   );
-  transition: left .9s ease;
+  transition: left 0.9s ease;
 }
 
 .submit-button:hover::before {
