+++ src/components/shared/CursorEffect/CustomCursor.tsx
@@ -1,0 +1,36 @@
+import { useState, useEffect } from 'react';
+import './CursorEffect.css';
+
+const CustomCursor = () => {
+  const [position, setPosition] = useState({ x: 0, y: 0 });
+  const [isPointer, setIsPointer] = useState(false);
+
+  useEffect(() => {
+    const moveCursor = (e: MouseEvent) => {
+      setPosition({ x: e.clientX, y: e.clientY });
+      
+      const target = e.target as HTMLElement;
+      setIsPointer(
+        window.getComputedStyle(target).getPropertyValue('cursor') === 'pointer'
+      );
+    };
+
+    window.addEventListener('mousemove', moveCursor);
+
+    return () => {
+      window.removeEventListener('mousemove', moveCursor);
+    };
+  }, []);
+
+  return (
+    <div 
+      className={`cursor ${isPointer ? 'cursor-pointer' : ''}`}
+      style={{
+        left: `${position.x}px`,
+        top: `${position.y}px`,
+      }}
+    />
+  );
+};
+
+export default CustomCursor;
