+++ src/components/shared/ErrorBoundary/ErrorBoundary.css
@@ -15,7 +15,7 @@
     padding: var(--spacing-xl);
     background-color: var(--bg-secondary);
     border-radius: var(--border-radius-lg);
-    box-shadow: 0 4px 12px var(--shadow-color);
+    box-shadow: 0 4px 2px var(--shadow-color);
     animation: fadeIn 0.5s ease;
   }
   
@@ -43,11 +43,6 @@
   }
   
   .error-reset-button:hover {
-    transform: translateY(-2px);
+    transform: scale(1.02);
   }
-  
-  .error-reset-button:active {
-    transform: translateY(0);
-  }
-
