+++ src/components/shared/ThemeToggle/ThemeToggle.tsx
@@ -11,7 +11,6 @@ const ThemeToggle: React.FC = () => {
   const [isDarkMode, setIsDarkMode] = useState(false)
 
   useEffect(() => {
-    // Check for saved theme preference or use system preference
     const savedTheme = localStorage.getItem("theme")
     const prefersDark = window.matchMedia("(prefers-color-scheme: dark)").matches
