+++ src/components/shared/icons/RefreshIcon.tsx
@@ -1,24 +1,25 @@
-import type React from "react"
+import type React from "react";
+import "./icons.css"; // Create this CSS file
+
 const RefreshIcon: React.FC = () => {
   return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="18"
-      height="18"
-      viewBox="0 0 24 24"
-      fill="none"
-      stroke="currentColor"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    >
-      <path d="M23 4v6h-6"></path>
-      <path d="M1 20v-6h6"></path>
-      <path d="M3.51 9a9 9 0 0 1 14.85-3.36L23 10"></path>
-      <path d="M1 14l4.64 4.36A9 9 0 0 0 20.49 15"></path>
-    </svg>
-  )
-}
+    <div className="refresh-icon">
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        viewBox="0 0 24 24"
+        fill="none"
+        className="refresh-svg"
+      >
+        <path
+          stroke="currentColor"
+          strokeWidth="2"
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          d="M21 12a9 9 0 1 1-2.636-6.364M21 4v4h-4"
+        />
+      </svg>
+    </div>
+  );
+};
 
-export default RefreshIcon
-
+export default RefreshIcon;
